Refatoração das entidades
