+++ src/main/java/br/com/zup/treinocasadocodigo/controllers/EstadoConrtoller.java
@@ -1,7 +1,7 @@
 package br.com.zup.treinocasadocodigo.controllers;
 
-import br.com.zup.treinocasadocodigo.entities.Estado;
-import br.com.zup.treinocasadocodigo.entities.EstadoNovoRequest;
+import br.com.zup.treinocasadocodigo.entities.estado.Estado;
+import br.com.zup.treinocasadocodigo.entities.estado.EstadoNovoRequest;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -10,6 +10,10 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+
+/**
+ * Contagem de carga intrínseca da classe: 1
+ */
 
 @RestController
 public class EstadoConrtoller {
