+++ src/main/java/br/com/zup/treinocasadocodigo/controllers/PaisController.java
@@ -1,9 +1,7 @@
 package br.com.zup.treinocasadocodigo.controllers;
 
-import br.com.zup.treinocasadocodigo.entities.Estado;
-import br.com.zup.treinocasadocodigo.entities.EstadoNovoRequest;
-import br.com.zup.treinocasadocodigo.entities.Pais;
-import br.com.zup.treinocasadocodigo.entities.PaisNovoRequest;
+import br.com.zup.treinocasadocodigo.entities.pais.Pais;
+import br.com.zup.treinocasadocodigo.entities.pais.PaisNovoRequest;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,6 +10,10 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+
+/**
+ * Contagem de carga intrínseca da classe: 1
+ */
 
 @RestController
 public class PaisController {
