+++ src/main/java/br/com/zup/treinocasadocodigo/entities/autor/AutorRetorno.java
@@ -1,4 +1,4 @@
-package br.com.zup.treinocasadocodigo.entities;
+package br.com.zup.treinocasadocodigo.entities.autor;
 
 public class AutorRetorno {
