+++ src/main/java/br/com/zup/treinocasadocodigo/entities/livro/LivroNovoRequest.java
@@ -1,5 +1,7 @@
-package br.com.zup.treinocasadocodigo.entities;
+package br.com.zup.treinocasadocodigo.entities.livro;
 
+import br.com.zup.treinocasadocodigo.entities.autor.Autor;
+import br.com.zup.treinocasadocodigo.entities.categoria.Categoria;
 import br.com.zup.treinocasadocodigo.validators.ExistId;
 import br.com.zup.treinocasadocodigo.validators.UniqueValue;
 import com.fasterxml.jackson.annotation.JsonFormat;
