+++ src/main/java/br/com/zup/treinocasadocodigo/entities/livro/LivroRetornoLista.java
@@ -1,4 +1,6 @@
-package br.com.zup.treinocasadocodigo.entities;
+package br.com.zup.treinocasadocodigo.entities.livro;
+
+import br.com.zup.treinocasadocodigo.entities.livro.Livro;
 
 /**
  * Contagem de carga intrínseca da classe: 1
